Add per-owner totals, multi-currency, and household scope
Three features that landed on top of the last commit. Functionally
separable but share many of the same files, so they ship together.

Per-owner monthly spent totals (Settings)
- AppState.monthlySpent(by:) aggregates expense rows in the current
  calendar month weighted by Transaction.effectiveSplits.
- Settings rows show "$X.XX this month" under each member, computed
  inline by SettingsView. User.detail field removed (no longer
  storage; the row's detail is derived).

Multi-currency with live FX rates
- Transaction.amount is now Int64 USD cents (was Decimal). All sums
  and call sites updated; signedAmount removed in favor of a
  leadingPlus flag on Money.string.
- New Models/Currency.swift — 7 currencies (USD, CAD, GBP, EUR, JPY,
  CNY, BDT) with ISO codes, display names, fraction digits, and
  hardcoded fallback rates.
- DesignSystem/Money.swift rewritten: string(cents:currency:rate:),
  symbol(for:), toUSDCents, toDisplayAmount. Cached per-currency
  NumberFormatter.
- AppState gains currency, fxRates, ratesLastFetched/IsLoading/Error.
  Live rates fetched on launch from floatrates.com (24h cache in
  UserDefaults, MainActor mutations, fallback when offline). New
  formatMoney(_:leadingPlus:) convenience used by every display site.
- RootTabView triggers refreshRatesIfStale() via .task.
- Settings gains a Currency Menu picker + rates-freshness caption.
- AddTransactionSheet hero glyph is currency-aware; submit converts
  user-typed amount → USD cents; edit mode pre-fills via toDisplay-
  Amount and preserves the stored cents when amountText is unchanged
  (originalAmountText round-trip mitigation).

Household functionality (Iteration 1)
- New Models/Household.swift — id, name, memberIDs + homeSample with
  stable UUID seeded with Maya + Jordan.
- Transaction.householdID: Household.ID? — nil = personal (single
  owner, visible only to that user); non-nil = shared with that
  household. Sample data tagged with homeSample.id.
- AppState: households, currentUserID, currentHouseholdID (both
  persisted via UserDefaults). Helpers currentHousehold,
  updateHouseholdName, addMemberToCurrentHousehold.
- Settings: USERS section replaced with HOUSEHOLD. Editable Name row
  (tap → SwiftUI .alert with TextField), "I am" Menu picker, member
  list with "(you)" indicator on the current user via the new
  UserRowView.isCurrentUser parameter.
- AddTransactionSheet: Shared | Personal segmented control above the
  Expense | Income toggle. Defaults to Shared. Personal collapses
  owners to {currentUserID} and hides the split editor.
- TransactionsView: All | Shared | Personal filter pill above the
  search field. Composes (AND) with merchant search.
- TransactionDetailSheet: nav title surfaces scope —
  "Expense · Home" / "Personal expense" / etc.

Out of scope (deferred):
- Transfers between members (Iteration 2).
- Multi-household UI / switcher (Iteration 3).
- Backend, auth, real invitations, multi-device sync.

+++ Ortho-iOS/Models/TransactionGroup.swift
@@ -16,7 +16,8 @@ struct TransactionGroup: Identifiable, Hashable {
         self.items = items
     }
 
-    var outgoingTotal: Decimal {
+    /// Sum (USD cents) of all expense rows in this day's bucket.
+    var outgoingTotal: Int64 {
         items.filter { $0.kind == .expense }.reduce(0) { $0 + $1.amount }
     }
